models/ai4robotics-wrist-movement-classification/requirements.txt
grpcio
grpc-interceptor
multithreading
protobuf
numpy
scikit-learn